messenger_service/requirements.txt
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.8.0
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
constantly==23.10.4
cryptography==46.0.5
daphne==4.2.1
#Django==6.0.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
packaging==26.0
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-dotenv==1.2.2
redis==7.2.1
service-identity==24.2.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
zope.interface==8.2
